Search for Wetlab Samples in the EPPIcenter SampleDB database

Search for wetlab samples.

Usage

SearchByType(
  sample_storage_type = NULL,
  control_type = NULL,
  filters = NULL,
  format = NULL,
  database = Sys.getenv("SDB_PATH"),
  config_yml = Sys.getenv("SDB_CONFIG"),
  include_internal_sample_id = FALSE
)

Arguments

sample_type

A string specifying the type of EPPIcenter sample. (micronix, cryovile, rdt and/or paper)

sample_label

A list of vectors specifying the vector micronix, cryovile, rdt, and paper label names (micronix.label_name, cryovile.label_name, rdt.label_name and/or paper.label_name)

container_name

A list of vectors specifying the vector micronix, cryovile, rdt, and paper container names (micronix.container_name, cryovile.container_name, rdt.container_name and/or paper.container_name)

study_subject

A study subjects string or a vector of study name strings. If study_subject.file is TRUE the path to a .csv file containing one column named study_subject can be uploaded and used to search the database for study subjects.

specimen_type

A specimen type string or a vector of specimen type strings.

study

A study short code string or a vector of study short code strings.

collection_dates

A list of date values strings (date.to and date.from) that can be used to filter EPPIcenter samples

archived

A logical value. TRUE filters for archived samples and FALSE filters for active samples

freezer

A list specifying the vector name, level_I, and/orlevel_II

return_sample_ids

A logical value. Setting return_sample_ids to TRUE means SearchSamples returns sample ids as well as search results. Setting return_sample_ids to FALSE means SearchSamples returns only search results. Default value is FALSE.

Examples

if (FALSE) { # \dontrun{
SearchSamples(study = "KAM06", study_subject = "subject_1")
} # }