Search for Wetlab Samples in the EPPIcenter SampleDB database

Usage

SearchSamples(
  sample_type = NULL,
  sample_label = NULL,
  container_name = NULL,
  study_subject = NULL,
  specimen_type = NULL,
  study = NULL,
  collection_dates = NULL,
  archived = NULL,
  freezer = NULL,
  study_subject.file = FALSE,
  return_sample_ids = FALSE
)

Arguments

sample_type

A string specifying the type of EPPIcenter sample. (micronix, cryovile, rdt and/or paper)

sample_label

A list of vectors specifying the vector micronix, cryovile, rdt, and paper label names (micronix.label_name, cryovile.label_name, rdt.label_name and/or paper.label_name)

container_name

A list of vectors specifying the vector micronix, cryovile, rdt, and paper container names (micronix.container_name, cryovile.container_name, rdt.container_name and/or paper.container_name)

study_subject

A study subjects string or a vector of study subject strings. If study_subject.file is TRUE the path to a .csv file containing one column named study_subject can be uploaded and used to search the database for study subjects.

specimen_type

A specimen type string or a vector of specimen type strings.

study

A study short code string or a vector of study short code strings.

collection_dates

A list of date values strings (date.to and date.from) that can be used to filter EPPIcenter samples

archived

A logical value. TRUE filters for archived samples and FALSE filters for active samples

freezer

A list specifying the vector location_name, level_I, and/orlevel_II

study_subject.file

A logical value. Setting study_subject.file to TRUE allows the user to search for study subjects using a .csv file. Setting study_subject.file to FALSE specifies that the user searches for study subjects using either a string or vector of strings. Default value is FALSE.

return_sample_ids

A logical value. Setting return_sample_ids to TRUE means SearchSamples returns sample ids as well as search results. Setting return_sample_ids to FALSE means SearchSamples returns only search results. Default value is FALSE.

Examples

if (FALSE) {
SearchSamples(study = "KAM06", study_subject = "subject_1")
}