Upload EPPIcenter Wetlab Samples to the SampleDB database

UploadSamples() can be used to upload wetlab samples to the sampleDB database. Currently the type of wetlab samples supported is only micronix. <!---cryovial`, `rdt` and `paper` sample uploads will appear in the next version.-->

Usage

UploadSamples(
  sample_type,
  upload_data,
  container_name,
  container_barcode = NULL,
  freezer_address
)

Arguments

sample_type

A string specifying the type of samples that are being uploaded Options include: micronix, cryovial, rdt and paper

upload_data

A dataframe of SampleDB Upload data.
Required upload_data columns are:
well_position: the row and column of the sample in the storage housing label: the sample's label or barcode study_subject_id: the participant id of for the subject in the cohort (ie study) study_short_code: the code of the study specimen_type: the sample type collection_date: (optional) the date the sample was first collected from the cohort participant

#' upload data example without collection_date

well_position	label	study_subject_id	specimen_type	study_short_code

A0	xxx1	subject_1	PLASMA	KAM06

A1	xxx2	subject_2	PLASMA	KAM06

upload data example with collection_date

well_position	label	study_subject_id	specimen_type	study_short_code	collection_data

A0	xxx1	subject_1	PLASMA	KAM06	2022-04-11

A1	xxx2	subject_2	PLASMA	KAM06	2022-04-11

container_name

A string specifying the name of the container the samples are in. Names must be unique within each sample type.

container_barcode

A string specifying the barcode for the container the samples are in. Container barcodes are optional. Barcodes must be unique within each sample type.

freezer_address

A list specifying the freezer address used to store samples.
Required items in the freezer_address list are location_name, level_I and level_II. If the freezer_address type is minus eighty then level_I and level_II items specify the rack and position, respecively. If the freezer_address type is minus twenty then level_I and level_II items specify the shelf and basket, respecively.

Examples

if (FALSE) {
UploadSamples(sample_type = "micronix", 
              upload_data = tibble(well_position = c("A0"),
                                   label = c("XXX 1"),
                                   study_subject_id = c("1"),
                                   specimen_type = c("PLASMA"),
                                   study_short_code = c("KAM06"),
                                   collection_date = c("2021-04-10")),
               container_name = "test_container", 
               freezer_address = list(location_name = "TBD",
                                      level_I = "TBD",
                                      level_II = "seve's working basket"))
}