Create Control Label with Density and Composition Values

This function takes vectors for density and composition and creates a control label by combining the values. Densities greater than 999 are divided by 1000 and appended with a 'K'.

Usage

create_control_label(dens_values, comp_values)

Arguments

dens_values

A numeric vector specifying the density values.

comp_values

A character vector specifying the composition values.

Value

A character vector with the control labels.

Examples

dens <- c(100, 1000, 1500)
comp <- c("S1_1", "S3_2", "S2_2")
create_control_label(dens, comp)
#> [1] "100_S1_1"  "1K_S3_2"   "1.5K_S2_2"