Set a Specific Row as the Header

This function updates the header of the user file based on the specified row number.

Usage

set_header_row(user_file, header_row)

Arguments

user_file

A data frame representing the user file.

header_row

A numeric value indicating which row to set as the header.

Value

A data frame with the updated header.

Examples

if (FALSE) { # \dontrun{
user_file <- read.csv("path_to_user_file.csv", header = FALSE)
set_header_row(user_file, 2)
} # }